Refactor main Spendvelope.java
