+++ README.md
@@ -1,4 +1,4 @@
-# Duke project template
+# Spendvelope project template
 
 This is a project template for a greenfield Java project. It's named after the Java mascot _Duke_. Given below are instructions on how to use it.
 
@@ -9,13 +9,13 @@ Prerequisites: JDK 11 (use the exact version), update Intellij to the most recen
 1. **Ensure Intellij JDK 11 is defined as an SDK**, as described [here](https://www.jetbrains.com/help/idea/sdk.html#set-up-jdk) -- this step is not needed if you have used JDK 11 in a previous Intellij project.
    * In the same dialog, you _may have to set the Project language level_ field to the SDK default option.
 2. **Import the project _as a Gradle project_**, as described [here](https://se-education.org/guides/tutorials/intellijImportGradleProject.html).
-3. **Verify the set up**: After the importing is complete, locate the `src/main/java/seedu/duke/Duke.java` file, right-click it, and choose `Run Duke.main()`. If the setup is correct, you should see something like the below:
+3. **Verify the set up**: After the importing is complete, locate the `src/main/java/seedu/duke/Spendvelope.java` file, right-click it, and choose `Run Spendvelope.main()`. If the setup is correct, you should see something like the below:
    ```
    > Task :compileJava
    > Task :processResources NO-SOURCE
    > Task :classes
    
-   > Task :Duke.main()
+   > Task :Spendvelope.main()
    Hello from
     ____        _        
    |  _ \ _   _| | _____
