+++ src/test/java/main/DukeTest.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package main;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
